[eBPF] Fixed a memory leak during uprobes processing #20814
A stable running Golang program (which has been running for at least 60 seconds) was probed and tracked by an agent using eBPF. If this Golang program exits, it will cause a memory leak. The reason for the memory leak is that when releasing the memory of the 'proc_info' (process information) structure during program exit, the memory occupied by the 'path' member (uprobe file path) inside the structure is not released, resulting in an average of about 60 bytes of memory leak. This patch fixes this bug.